ANTIC-001 After the descussion with the group the register logic and local db is not needed because of firebase, but it may be relevant in the future. The updates in this commit are the ui updates that made the register ui ready [BL, BC]
