--- AntiCovid/app/src/main/java/com/swtug/anticovid/ui/login/LoginFormState.kt
@@ -1,10 +1,0 @@
-package com.swtug.anticovid.ui.login
-
-/**
- * Data validation state of the login form.
- */
-data class LoginFormState(
-    val usernameError: Int? = null,
-    val passwordError: Int? = null,
-    val isDataValid: Boolean = false
-)
