+++ AntiCovid/app/src/main/res/values/strings.xml
@@ -8,7 +8,7 @@
     <!-- Strings related to login -->
     <string name="prompt_email">Email</string>
     <string name="prompt_password">Password</string>
-    <string name="action_sign_in">Sign up</string>
+    <string name="action_sign_in">Sign up -></string>
     <string name="action_sign_in_short">Sign in</string>
     <string name="welcome">"Welcome!"</string>
     <string name="invalid_username">Not a valid username</string>
@@ -20,4 +20,5 @@
     <string name="phone_number">Phone number</string>
     <string name="name">Name</string>
     <string name="surname">Surname</string>
+    <string name="register_title">Create Account</string>
 </resources>
